feat(71-01): migrate StatsIntervalSeconds from int to double with validation
Data layer changes:
- AppSettings.StatsIntervalSeconds: int = 3 → double = 2.0
- SettingsSnapshot.StatsIntervalSeconds: int → double
- SettingsService.Validate(): range check 0.5-10.0, Math.Round(value, 1) precision
- SettingsService.Defaults(): StatsIntervalSeconds = 2.0

MainWindow minimal changes (enable tests to compile):
- _statsIntervalSeconds field: int = 3 → double = 2.0
- SetStatsInterval(double) signature change
- maxSamples division: explicit (int) cast

Tests (TDD RED→GREEN):
- 6 new SettingsServiceTests: BelowMin, AboveMax, ValidValue, RoundsPrecision, Defaults_Is2
- Update existing Validate_ZeroStatsInterval to expect 2.0
- AppSettingsTests: RoundTrip uses 2.5 with epsilon, Deserialize_IntegerStatsInterval migration test
- All 49 SettingsService/AppSettings tests pass

+++ FuzzyClock.App/SettingsSnapshot.cs
@@ -27,7 +27,7 @@ internal sealed record SettingsSnapshot
     public bool    PagVisible                            { get; init; }
     public bool    BatteryVisible                        { get; init; }
     public bool    UptimeVisible                         { get; init; }
-    public int     StatsIntervalSeconds                  { get; init; }
+    public double  StatsIntervalSeconds                  { get; init; }
     public double  ProcessCountThreshold                 { get; init; }
     public bool    ShowDate                              { get; init; }
     public string  DateFormat                            { get; init; } = "Short";
